batch/requirements.txt
pandas
psycopg2-binary
geopandas
sqlalchemy
geoalchemy2
